Add session check for simultaneous use
Get the session count which is defined the simultaneous use,
FreeRadius will use the session count to compare with maximum setting.

The maximum setting is defined by "Simultaneous-Use" attribute in
users/groups check collection per group or per user.

The user authorization request is rejected if the session count reaches
the maximum setting.

Signed-off-by: Neutron Soutmun <neo.neutron@gmail.com>