Add verify and zap stale sessions
Add verify session to query all sessions of requesting user and calls external
program that defined by FreeRADIUS configuration "checkrad" to verify the stale
session and zap it.

This could be disabled by set the configuration "verify_session" to "no" and
set "zap_stale_sessions" to "no" if you do not want to zap the stale
sessions. The "zap_stale_sessions" is ignored if the "verify_session" is
disabled.

Adjust the accounting type "stop" which the "Acct-Terminate-Cause" is
not fundamental anymore, just update if it's exists. (The zap call has
not set any "Acct-Terminate-Cause")

Signed-off-by: Neutron Soutmun <neo.neutron@gmail.com>